Old Version/requirements.txt
colorthief
Pillow
requests
